backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn>=0.29.0,<1.0.0
boto3>=1.34.0,<2.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
hypothesis>=6.98.0,<7.0.0
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<1.0.0
pyyaml>=6.0,<7.0.0
